edited controller and service implementation
as per dion's prior feedback, i edited the methods such that there is a separation of concerns. Controller interacts with the service classes and not the repository. Also fixed the addtoCart method which was logically incorrect but should be correct now. Controller now uses methods instead of business logic code. Most business logic confined to service class . @Montichiari please help me take a look if it is in order. I also need your help to add the files that may be missing since we discussed last night that github is incorrectly syncing the changes u made with my branch. Thanks !